// src/feutures/EditableProfileCard/model/const/editableProfileCardConsts.ts
export enum ValidateErrors {
    INCORRECT_FIRSTNAME = 'INCORRECT_FIRSTNAME',
    INCORRECT_LASTNAME = 'INCORRECT_LASTNAME',
    INCORRECT_AGE = 'INCORRECT_AGE',
    INCORRECT_CITY = 'INCORRECT_CITY',
    INCORRECT_USERNAME = 'INCORRECT_USERNAME',
    INCORRECT_AVATAR = 'INCORRECT_AVATAR',
    INCORRECT_DATA = 'INCORRECT_DATA',
    SERVER_ERROR = 'SERVER_ERROR'
}